fix, feat, refactor: 뽑기 카드 크기 원상복구, 카드 섞는 함수 분리, type들 interface로 변경, Suspense fallback 로딩 컴포넌트로 수정
