+++ src/app/draw/result/DrawList.tsx
@@ -4,6 +4,7 @@ import { useEffect, useRef, useState } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faRefresh } from '@fortawesome/free-solid-svg-icons';
 import Button from '@/components/Button';
+import { mixCards } from '@/utils/draw';
 
 export default function DrawList({ total, win }: DrawListProps) {
   const [cards, setCards] = useState<DrawItem[]>([]);
@@ -23,7 +24,7 @@ export default function DrawList({ total, win }: DrawListProps) {
     <div className="flex ">
       {cards.map((card, index) => (
         <button
-          className="absolute transition-all duration-300 ease-in-out"
+          className="w-50 absolute transition-all duration-300 ease-in-out"
           key={card.num}
           style={{
             left: `calc(50% + ${-400 / 2 + card.x * width}px)`,
@@ -80,24 +81,4 @@ export default function DrawList({ total, win }: DrawListProps) {
       )}
     </div>
   );
-}
-
-function getRandomHexColor(): string {
-  const randomColor = Math.floor(Math.random() * 16777215).toString(16);
-  return `#${randomColor.padStart(6, '0')}`;
-}
-
-function mixCards(total: number, win: number, cards?: DrawItem[]): DrawItem[] {
-  const arr: number[] = Array.from(new Array(total), (_, i) => i);
-  arr.sort(() => Math.random() - 0.5);
-  return Array.from(cards || new Array(total), (_, i) => {
-    return {
-      x: Math.random(),
-      y: Math.random(),
-      color: getRandomHexColor(),
-      num: cards ? cards[i].num : arr[i],
-      angle: Math.random() * 360,
-      isWinner: cards ? cards[i].isWinner : arr[i] >= total - win,
-    };
-  });
 }
