+++ tsconfig.json
@@ -19,8 +19,7 @@
       }
     ],
     "paths": {
-      "@/*": ["./src/*"],
-      "types": ["./src/types"]
+      "@/*": ["./src/*"]
     },
     "typeRoots": ["node_modules/@types", "./src/types"]
   },
